Map Explanation
The files I changed were the version 5 files. I added a map to the home page where you can click different points on the map. Once you click, you'll be taken to a specific point on the page. This helped me understand how to use maps, create interactive images and how to bring the user to different places on the page with <a> and href