FFT
I created an image holder that held images in rows that adjusted to the images. I added this to my index file in Version 5.0 with the rest of my other FFT projects. This helped me better understand HTML and CSS because it helped me understand flex display, flex-wrap and justify-content.